+++ Sources/PCAttachment.swift
@@ -1,0 +1,8 @@
+import Foundation
+import PusherPlatform
+
+public struct PCAttachment {
+    public let fetchRequired: Bool
+    public let link: String
+    public let type: String
+}