+++ Sources/PCAttachmentUploadResponse.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct PCAttachmentUploadResponse {
+    public let link: String
+    public let type: String
+}